Crate rusoto_servicecatalog

AWS Service Catalog
AWS Service Catalog enables organizations to create and manage catalogs of IT services that are approved for use on AWS. To get the most out of this documentation, you should be familiar with the terminology discussed in AWS Service Catalog Concepts.

If you're using the service, you're probably looking for ServiceCatalogClient and ServiceCatalog.

Structs

AcceptPortfolioShareInput

AcceptPortfolioShareOutput

AccessLevelFilter
The access level to use to filter results.

AssociateBudgetWithResourceInput

AssociateBudgetWithResourceOutput

AssociatePrincipalWithPortfolioInput

AssociatePrincipalWithPortfolioOutput

AssociateProductWithPortfolioInput

AssociateProductWithPortfolioOutput

AssociateServiceActionWithProvisioningArtifactInput

AssociateServiceActionWithProvisioningArtifactOutput

AssociateTagOptionWithResourceInput

AssociateTagOptionWithResourceOutput

BatchAssociateServiceActionWithProvisioningArtifactInput

BatchAssociateServiceActionWithProvisioningArtifactOutput

BatchDisassociateServiceActionFromProvisioningArtifactInput

BatchDisassociateServiceActionFromProvisioningArtifactOutput

BudgetDetail
Information about a budget.

CloudWatchDashboard
Information about a CloudWatch dashboard.

ConstraintDetail
Information about a constraint.

ConstraintSummary
Summary information about a constraint.

CopyProductInput

CopyProductOutput

CreateConstraintInput

CreateConstraintOutput

CreatePortfolioInput

CreatePortfolioOutput

CreatePortfolioShareInput

CreatePortfolioShareOutput

CreateProductInput

CreateProductOutput

CreateProvisionedProductPlanInput

CreateProvisionedProductPlanOutput

CreateProvisioningArtifactInput

CreateProvisioningArtifactOutput

CreateServiceActionInput

CreateServiceActionOutput

CreateTagOptionInput

CreateTagOptionOutput

DeleteConstraintInput

DeleteConstraintOutput

DeletePortfolioInput

DeletePortfolioOutput

DeletePortfolioShareInput

DeletePortfolioShareOutput

DeleteProductInput

DeleteProductOutput

DeleteProvisionedProductPlanInput

DeleteProvisionedProductPlanOutput

DeleteProvisioningArtifactInput

DeleteProvisioningArtifactOutput

DeleteServiceActionInput

DeleteServiceActionOutput

DeleteTagOptionInput

DeleteTagOptionOutput

DescribeConstraintInput

DescribeConstraintOutput

DescribeCopyProductStatusInput

DescribeCopyProductStatusOutput

DescribePortfolioInput

DescribePortfolioOutput

DescribePortfolioShareStatusInput

DescribePortfolioShareStatusOutput

DescribeProductAsAdminInput

DescribeProductAsAdminOutput

DescribeProductInput

DescribeProductOutput

DescribeProductViewInput

DescribeProductViewOutput

DescribeProvisionedProductInput

DescribeProvisionedProductOutput

DescribeProvisionedProductPlanInput

DescribeProvisionedProductPlanOutput

DescribeProvisioningArtifactInput

DescribeProvisioningArtifactOutput

DescribeProvisioningParametersInput

DescribeProvisioningParametersOutput

DescribeRecordInput

DescribeRecordOutput

DescribeServiceActionInput

DescribeServiceActionOutput

DescribeTagOptionInput

DescribeTagOptionOutput

DisableAWSOrganizationsAccessInput

DisableAWSOrganizationsAccessOutput

DisassociateBudgetFromResourceInput

DisassociateBudgetFromResourceOutput

DisassociatePrincipalFromPortfolioInput

DisassociatePrincipalFromPortfolioOutput

DisassociateProductFromPortfolioInput

DisassociateProductFromPortfolioOutput

DisassociateServiceActionFromProvisioningArtifactInput

DisassociateServiceActionFromProvisioningArtifactOutput

DisassociateTagOptionFromResourceInput

DisassociateTagOptionFromResourceOutput

EnableAWSOrganizationsAccessInput

EnableAWSOrganizationsAccessOutput

ExecuteProvisionedProductPlanInput

ExecuteProvisionedProductPlanOutput

ExecuteProvisionedProductServiceActionInput

ExecuteProvisionedProductServiceActionOutput

FailedServiceActionAssociation
An object containing information about the error, along with identifying information about the self-service action and its associations.

GetAWSOrganizationsAccessStatusInput

GetAWSOrganizationsAccessStatusOutput

LaunchPathSummary
Summary information about a product path for a user.

ListAcceptedPortfolioSharesInput

ListAcceptedPortfolioSharesOutput

ListBudgetsForResourceInput

ListBudgetsForResourceOutput

ListConstraintsForPortfolioInput

ListConstraintsForPortfolioOutput

ListLaunchPathsInput

ListLaunchPathsOutput

ListOrganizationPortfolioAccessInput

ListOrganizationPortfolioAccessOutput

ListPortfolioAccessInput

ListPortfolioAccessOutput

ListPortfoliosForProductInput

ListPortfoliosForProductOutput

ListPortfoliosInput

ListPortfoliosOutput

ListPrincipalsForPortfolioInput

ListPrincipalsForPortfolioOutput

ListProvisionedProductPlansInput

ListProvisionedProductPlansOutput

ListProvisioningArtifactsForServiceActionInput

ListProvisioningArtifactsForServiceActionOutput

ListProvisioningArtifactsInput

ListProvisioningArtifactsOutput

ListRecordHistoryInput

ListRecordHistoryOutput

ListRecordHistorySearchFilter
The search filter to use when listing history records.

ListResourcesForTagOptionInput

ListResourcesForTagOptionOutput

ListServiceActionsForProvisioningArtifactInput

ListServiceActionsForProvisioningArtifactOutput

ListServiceActionsInput

ListServiceActionsOutput

ListStackInstancesForProvisionedProductInput

ListStackInstancesForProvisionedProductOutput

ListTagOptionsFilters
Filters to use when listing TagOptions.

ListTagOptionsInput

ListTagOptionsOutput

OrganizationNode
Information about the organization node.

ParameterConstraints
The constraints that the administrator has put on the parameter.

PortfolioDetail
Information about a portfolio.

Principal
Information about a principal.

ProductViewAggregationValue
A single product view aggregation value/count pair, containing metadata about each product to which the calling user has access.

ProductViewDetail
Information about a product view.

ProductViewSummary
Summary information about a product view.

ProvisionProductInput

ProvisionProductOutput

ProvisionedProductAttribute
Information about a provisioned product.

ProvisionedProductDetail
Information about a provisioned product.

ProvisionedProductPlanDetails
Information about a plan.

ProvisionedProductPlanSummary
Summary information about a plan.

ProvisioningArtifact
Information about a provisioning artifact. A provisioning artifact is also known as a product version.

ProvisioningArtifactDetail
Information about a provisioning artifact (also known as a version) for a product.

ProvisioningArtifactParameter
Information about a parameter used to provision a product.

ProvisioningArtifactPreferences
The user-defined preferences that will be applied during product provisioning, unless overridden by ProvisioningPreferences or UpdateProvisioningPreferences.

For more information on maximum concurrent accounts and failure tolerance, see Stack set operation options in the AWS CloudFormation User Guide.

ProvisioningArtifactProperties
Information about a provisioning artifact (also known as a version) for a product.

ProvisioningArtifactSummary
Summary information about a provisioning artifact (also known as a version) for a product.

ProvisioningArtifactView
An object that contains summary information about a product view and a provisioning artifact.

ProvisioningParameter
Information about a parameter used to provision a product.

ProvisioningPreferences
The user-defined preferences that will be applied when updating a provisioned product. Not all preferences are applicable to all provisioned product types.

RecordDetail
Information about a request operation.

RecordError
The error code and description resulting from an operation.

RecordOutput
The output for the product created as the result of a request. For example, the output for a CloudFormation-backed product that creates an S3 bucket would include the S3 bucket URL.

RecordTag
Information about a tag, which is a key-value pair.

RejectPortfolioShareInput

RejectPortfolioShareOutput

ResourceChange
Information about a resource change that will occur when a plan is executed.

ResourceChangeDetail
Information about a change to a resource attribute.

ResourceDetail
Information about a resource.

ResourceTargetDefinition
Information about a change to a resource attribute.

ScanProvisionedProductsInput

ScanProvisionedProductsOutput

SearchProductsAsAdminInput

SearchProductsAsAdminOutput

SearchProductsInput

SearchProductsOutput

SearchProvisionedProductsInput

SearchProvisionedProductsOutput

ServiceActionAssociation
A self-service action association consisting of the Action ID, the Product ID, and the Provisioning Artifact ID.

ServiceActionDetail
An object containing detailed information about the self-service action.

ServiceActionSummary
Detailed information about the self-service action.

ServiceCatalogClient
A client for the AWS Service Catalog API.

ShareDetails
Information about the portfolio share operation.

ShareError
Errors that occurred during the portfolio share operation.

StackInstance
An AWS CloudFormation stack, in a specific account and region, that's part of a stack set operation. A stack instance is a reference to an attempted or actual stack in a given account within a given region. A stack instance can exist without a stack—for example, if the stack couldn't be created for some reason. A stack instance is associated with only one stack set. Each stack instance contains the ID of its associated stack set, as well as the ID of the actual stack and the stack status.

Tag
Information about a tag. A tag is a key-value pair. Tags are propagated to the resources created when provisioning a product.

TagOptionDetail
Information about a TagOption.

TagOptionSummary
Summary information about a TagOption.

TerminateProvisionedProductInput

TerminateProvisionedProductOutput

UpdateConstraintInput

UpdateConstraintOutput

UpdatePortfolioInput

UpdatePortfolioOutput

UpdateProductInput

UpdateProductOutput

UpdateProvisionedProductInput

UpdateProvisionedProductOutput

UpdateProvisionedProductPropertiesInput

UpdateProvisionedProductPropertiesOutput

UpdateProvisioningArtifactInput

UpdateProvisioningArtifactOutput

UpdateProvisioningParameter
The parameter key-value pair used to update a provisioned product.

UpdateProvisioningPreferences
The user-defined preferences that will be applied when updating a provisioned product. Not all preferences are applicable to all provisioned product types.

UpdateServiceActionInput

UpdateServiceActionOutput

UpdateTagOptionInput

UpdateTagOptionOutput

UsageInstruction
Additional information provided by the administrator.

Enums

AcceptPortfolioShareError
Errors returned by AcceptPortfolioShare

AssociateBudgetWithResourceError
Errors returned by AssociateBudgetWithResource

AssociatePrincipalWithPortfolioError
Errors returned by AssociatePrincipalWithPortfolio

AssociateProductWithPortfolioError
Errors returned by AssociateProductWithPortfolio

AssociateServiceActionWithProvisioningArtifactError
Errors returned by AssociateServiceActionWithProvisioningArtifact

AssociateTagOptionWithResourceError
Errors returned by AssociateTagOptionWithResource

BatchAssociateServiceActionWithProvisioningArtifactError
Errors returned by BatchAssociateServiceActionWithProvisioningArtifact

BatchDisassociateServiceActionFromProvisioningArtifactError
Errors returned by BatchDisassociateServiceActionFromProvisioningArtifact

CopyProductError
Errors returned by CopyProduct

CreateConstraintError
Errors returned by CreateConstraint

CreatePortfolioError
Errors returned by CreatePortfolio

CreatePortfolioShareError
Errors returned by CreatePortfolioShare

CreateProductError
Errors returned by CreateProduct

CreateProvisionedProductPlanError
Errors returned by CreateProvisionedProductPlan

CreateProvisioningArtifactError
Errors returned by CreateProvisioningArtifact

CreateServiceActionError
Errors returned by CreateServiceAction

CreateTagOptionError
Errors returned by CreateTagOption

DeleteConstraintError
Errors returned by DeleteConstraint

DeletePortfolioError
Errors returned by DeletePortfolio

DeletePortfolioShareError
Errors returned by DeletePortfolioShare

DeleteProductError
Errors returned by DeleteProduct

DeleteProvisionedProductPlanError
Errors returned by DeleteProvisionedProductPlan

DeleteProvisioningArtifactError
Errors returned by DeleteProvisioningArtifact

DeleteServiceActionError
Errors returned by DeleteServiceAction

DeleteTagOptionError
Errors returned by DeleteTagOption

DescribeConstraintError
Errors returned by DescribeConstraint

DescribeCopyProductStatusError
Errors returned by DescribeCopyProductStatus

DescribePortfolioError
Errors returned by DescribePortfolio

DescribePortfolioShareStatusError
Errors returned by DescribePortfolioShareStatus

DescribeProductAsAdminError
Errors returned by DescribeProductAsAdmin

DescribeProductError
Errors returned by DescribeProduct

DescribeProductViewError
Errors returned by DescribeProductView

DescribeProvisionedProductError
Errors returned by DescribeProvisionedProduct

DescribeProvisionedProductPlanError
Errors returned by DescribeProvisionedProductPlan

DescribeProvisioningArtifactError
Errors returned by DescribeProvisioningArtifact

DescribeProvisioningParametersError
Errors returned by DescribeProvisioningParameters

DescribeRecordError
Errors returned by DescribeRecord

DescribeServiceActionError
Errors returned by DescribeServiceAction

DescribeTagOptionError
Errors returned by DescribeTagOption

DisableAWSOrganizationsAccessError
Errors returned by DisableAWSOrganizationsAccess

DisassociateBudgetFromResourceError
Errors returned by DisassociateBudgetFromResource

DisassociatePrincipalFromPortfolioError
Errors returned by DisassociatePrincipalFromPortfolio

DisassociateProductFromPortfolioError
Errors returned by DisassociateProductFromPortfolio

DisassociateServiceActionFromProvisioningArtifactError
Errors returned by DisassociateServiceActionFromProvisioningArtifact

DisassociateTagOptionFromResourceError
Errors returned by DisassociateTagOptionFromResource

EnableAWSOrganizationsAccessError
Errors returned by EnableAWSOrganizationsAccess

ExecuteProvisionedProductPlanError
Errors returned by ExecuteProvisionedProductPlan

ExecuteProvisionedProductServiceActionError
Errors returned by ExecuteProvisionedProductServiceAction

GetAWSOrganizationsAccessStatusError
Errors returned by GetAWSOrganizationsAccessStatus

ListAcceptedPortfolioSharesError
Errors returned by ListAcceptedPortfolioShares

ListBudgetsForResourceError
Errors returned by ListBudgetsForResource

ListConstraintsForPortfolioError
Errors returned by ListConstraintsForPortfolio

ListLaunchPathsError
Errors returned by ListLaunchPaths

ListOrganizationPortfolioAccessError
Errors returned by ListOrganizationPortfolioAccess

ListPortfolioAccessError
Errors returned by ListPortfolioAccess

ListPortfoliosError
Errors returned by ListPortfolios

ListPortfoliosForProductError
Errors returned by ListPortfoliosForProduct

ListPrincipalsForPortfolioError
Errors returned by ListPrincipalsForPortfolio

ListProvisionedProductPlansError
Errors returned by ListProvisionedProductPlans

ListProvisioningArtifactsError
Errors returned by ListProvisioningArtifacts

ListProvisioningArtifactsForServiceActionError
Errors returned by ListProvisioningArtifactsForServiceAction

ListRecordHistoryError
Errors returned by ListRecordHistory

ListResourcesForTagOptionError
Errors returned by ListResourcesForTagOption

ListServiceActionsError
Errors returned by ListServiceActions

ListServiceActionsForProvisioningArtifactError
Errors returned by ListServiceActionsForProvisioningArtifact

ListStackInstancesForProvisionedProductError
Errors returned by ListStackInstancesForProvisionedProduct

ListTagOptionsError
Errors returned by ListTagOptions

ProvisionProductError
Errors returned by ProvisionProduct

RejectPortfolioShareError
Errors returned by RejectPortfolioShare

ScanProvisionedProductsError
Errors returned by ScanProvisionedProducts

SearchProductsAsAdminError
Errors returned by SearchProductsAsAdmin

SearchProductsError
Errors returned by SearchProducts

SearchProvisionedProductsError
Errors returned by SearchProvisionedProducts

TerminateProvisionedProductError
Errors returned by TerminateProvisionedProduct

UpdateConstraintError
Errors returned by UpdateConstraint

UpdatePortfolioError
Errors returned by UpdatePortfolio

UpdateProductError
Errors returned by UpdateProduct

UpdateProvisionedProductError
Errors returned by UpdateProvisionedProduct

UpdateProvisionedProductPropertiesError
Errors returned by UpdateProvisionedProductProperties

UpdateProvisioningArtifactError
Errors returned by UpdateProvisioningArtifact

UpdateServiceActionError
Errors returned by UpdateServiceAction

UpdateTagOptionError
Errors returned by UpdateTagOption

Traits

ServiceCatalog
Trait representing the capabilities of the AWS Service Catalog API. AWS Service Catalog clients implement this trait.