Struct rusoto_servicecatalog::ProvisionedProductDetail

pub struct ProvisionedProductDetail {
    pub arn: Option<String>,
    pub created_time: Option<f64>,
    pub id: Option<String>,
    pub idempotency_token: Option<String>,
    pub last_record_id: Option<String>,
    pub name: Option<String>,
    pub status: Option<String>,
    pub status_message: Option<String>,
    pub type_: Option<String>,
}

Detailed information about a ProvisionedProduct object.

Fields
arn: Option<String>

The ARN associated with the ProvisionedProduct object.

created_time: Option<f64>

The UTC timestamp of the creation time.

id: Option<String>

The identifier of the ProvisionedProduct object.

idempotency_token: Option<String>

A token to disambiguate duplicate requests. You can create multiple resources using the same input in multiple requests, provided that you also specify a different idempotency token for each request.

last_record_id: Option<String>

The record identifier of the last request performed on this ProvisionedProduct object.

name: Option<String>

The user-friendly name of the ProvisionedProduct object.

status: Option<String>

The current status of the ProvisionedProduct.

AVAILABLE - Stable state, ready to perform any operation. The most recent action request succeeded and completed.

UNDER_CHANGE - Transitive state, operations performed may or may not have valid results. Wait for an AVAILABLE status before performing operations.

TAINTED - Stable state, ready to perform any operation. The stack has completed the requested operation but is not exactly what was requested. For example, a request to update to a new version failed and the stack rolled back to the current version.

ERROR - Something unexpected happened such that the provisioned product exists but the stack is not running. For example, CloudFormation received an invalid parameter value and could not launch the stack.

status_message: Option<String>

The current status message of the ProvisionedProduct.

type_: Option<String>

The type of the ProvisionedProduct object.