Struct rusoto_servicecatalog::ServiceCatalogClient

pub struct ServiceCatalogClient { /* fields omitted */ }

A client for the AWS Service Catalog API.

Methods

impl ServiceCatalogClient

pub fn new(region: Region) -> ServiceCatalogClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> ServiceCatalogClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,