Trait rusoto_servicecatalog::ServiceCatalog

pub trait ServiceCatalog {
    fn accept_portfolio_share(
        &self, 
        input: &AcceptPortfolioShareInput
    ) -> Result<AcceptPortfolioShareOutput, AcceptPortfolioShareError>;
    fn associate_principal_with_portfolio(
        &self, 
        input: &AssociatePrincipalWithPortfolioInput
    ) -> Result<AssociatePrincipalWithPortfolioOutput, AssociatePrincipalWithPortfolioError>;
    fn associate_product_with_portfolio(
        &self, 
        input: &AssociateProductWithPortfolioInput
    ) -> Result<AssociateProductWithPortfolioOutput, AssociateProductWithPortfolioError>;
    fn create_constraint(
        &self, 
        input: &CreateConstraintInput
    ) -> Result<CreateConstraintOutput, CreateConstraintError>;
    fn create_portfolio(
        &self, 
        input: &CreatePortfolioInput
    ) -> Result<CreatePortfolioOutput, CreatePortfolioError>;
    fn create_portfolio_share(
        &self, 
        input: &CreatePortfolioShareInput
    ) -> Result<CreatePortfolioShareOutput, CreatePortfolioShareError>;
    fn create_product(
        &self, 
        input: &CreateProductInput
    ) -> Result<CreateProductOutput, CreateProductError>;
    fn create_provisioning_artifact(
        &self, 
        input: &CreateProvisioningArtifactInput
    ) -> Result<CreateProvisioningArtifactOutput, CreateProvisioningArtifactError>;
    fn delete_constraint(
        &self, 
        input: &DeleteConstraintInput
    ) -> Result<DeleteConstraintOutput, DeleteConstraintError>;
    fn delete_portfolio(
        &self, 
        input: &DeletePortfolioInput
    ) -> Result<DeletePortfolioOutput, DeletePortfolioError>;
    fn delete_portfolio_share(
        &self, 
        input: &DeletePortfolioShareInput
    ) -> Result<DeletePortfolioShareOutput, DeletePortfolioShareError>;
    fn delete_product(
        &self, 
        input: &DeleteProductInput
    ) -> Result<DeleteProductOutput, DeleteProductError>;
    fn delete_provisioning_artifact(
        &self, 
        input: &DeleteProvisioningArtifactInput
    ) -> Result<DeleteProvisioningArtifactOutput, DeleteProvisioningArtifactError>;
    fn describe_constraint(
        &self, 
        input: &DescribeConstraintInput
    ) -> Result<DescribeConstraintOutput, DescribeConstraintError>;
    fn describe_portfolio(
        &self, 
        input: &DescribePortfolioInput
    ) -> Result<DescribePortfolioOutput, DescribePortfolioError>;
    fn describe_product(
        &self, 
        input: &DescribeProductInput
    ) -> Result<DescribeProductOutput, DescribeProductError>;
    fn describe_product_as_admin(
        &self, 
        input: &DescribeProductAsAdminInput
    ) -> Result<DescribeProductAsAdminOutput, DescribeProductAsAdminError>;
    fn describe_product_view(
        &self, 
        input: &DescribeProductViewInput
    ) -> Result<DescribeProductViewOutput, DescribeProductViewError>;
    fn describe_provisioned_product(
        &self, 
        input: &DescribeProvisionedProductInput
    ) -> Result<DescribeProvisionedProductOutput, DescribeProvisionedProductError>;
    fn describe_provisioning_artifact(
        &self, 
        input: &DescribeProvisioningArtifactInput
    ) -> Result<DescribeProvisioningArtifactOutput, DescribeProvisioningArtifactError>;
    fn describe_provisioning_parameters(
        &self, 
        input: &DescribeProvisioningParametersInput
    ) -> Result<DescribeProvisioningParametersOutput, DescribeProvisioningParametersError>;
    fn describe_record(
        &self, 
        input: &DescribeRecordInput
    ) -> Result<DescribeRecordOutput, DescribeRecordError>;
    fn disassociate_principal_from_portfolio(
        &self, 
        input: &DisassociatePrincipalFromPortfolioInput
    ) -> Result<DisassociatePrincipalFromPortfolioOutput, DisassociatePrincipalFromPortfolioError>;
    fn disassociate_product_from_portfolio(
        &self, 
        input: &DisassociateProductFromPortfolioInput
    ) -> Result<DisassociateProductFromPortfolioOutput, DisassociateProductFromPortfolioError>;
    fn list_accepted_portfolio_shares(
        &self, 
        input: &ListAcceptedPortfolioSharesInput
    ) -> Result<ListAcceptedPortfolioSharesOutput, ListAcceptedPortfolioSharesError>;
    fn list_constraints_for_portfolio(
        &self, 
        input: &ListConstraintsForPortfolioInput
    ) -> Result<ListConstraintsForPortfolioOutput, ListConstraintsForPortfolioError>;
    fn list_launch_paths(
        &self, 
        input: &ListLaunchPathsInput
    ) -> Result<ListLaunchPathsOutput, ListLaunchPathsError>;
    fn list_portfolio_access(
        &self, 
        input: &ListPortfolioAccessInput
    ) -> Result<ListPortfolioAccessOutput, ListPortfolioAccessError>;
    fn list_portfolios(
        &self, 
        input: &ListPortfoliosInput
    ) -> Result<ListPortfoliosOutput, ListPortfoliosError>;
    fn list_portfolios_for_product(
        &self, 
        input: &ListPortfoliosForProductInput
    ) -> Result<ListPortfoliosForProductOutput, ListPortfoliosForProductError>;
    fn list_principals_for_portfolio(
        &self, 
        input: &ListPrincipalsForPortfolioInput
    ) -> Result<ListPrincipalsForPortfolioOutput, ListPrincipalsForPortfolioError>;
    fn list_provisioning_artifacts(
        &self, 
        input: &ListProvisioningArtifactsInput
    ) -> Result<ListProvisioningArtifactsOutput, ListProvisioningArtifactsError>;
    fn list_record_history(
        &self, 
        input: &ListRecordHistoryInput
    ) -> Result<ListRecordHistoryOutput, ListRecordHistoryError>;
    fn provision_product(
        &self, 
        input: &ProvisionProductInput
    ) -> Result<ProvisionProductOutput, ProvisionProductError>;
    fn reject_portfolio_share(
        &self, 
        input: &RejectPortfolioShareInput
    ) -> Result<RejectPortfolioShareOutput, RejectPortfolioShareError>;
    fn scan_provisioned_products(
        &self, 
        input: &ScanProvisionedProductsInput
    ) -> Result<ScanProvisionedProductsOutput, ScanProvisionedProductsError>;
    fn search_products(
        &self, 
        input: &SearchProductsInput
    ) -> Result<SearchProductsOutput, SearchProductsError>;
    fn search_products_as_admin(
        &self, 
        input: &SearchProductsAsAdminInput
    ) -> Result<SearchProductsAsAdminOutput, SearchProductsAsAdminError>;
    fn terminate_provisioned_product(
        &self, 
        input: &TerminateProvisionedProductInput
    ) -> Result<TerminateProvisionedProductOutput, TerminateProvisionedProductError>;
    fn update_constraint(
        &self, 
        input: &UpdateConstraintInput
    ) -> Result<UpdateConstraintOutput, UpdateConstraintError>;
    fn update_portfolio(
        &self, 
        input: &UpdatePortfolioInput
    ) -> Result<UpdatePortfolioOutput, UpdatePortfolioError>;
    fn update_product(
        &self, 
        input: &UpdateProductInput
    ) -> Result<UpdateProductOutput, UpdateProductError>;
    fn update_provisioned_product(
        &self, 
        input: &UpdateProvisionedProductInput
    ) -> Result<UpdateProvisionedProductOutput, UpdateProvisionedProductError>;
    fn update_provisioning_artifact(
        &self, 
        input: &UpdateProvisioningArtifactInput
    ) -> Result<UpdateProvisioningArtifactOutput, UpdateProvisioningArtifactError>;
}

Trait representing the capabilities of the AWS Service Catalog API. AWS Service Catalog clients implement this trait.

Required Methods

fn accept_portfolio_share(
&self,
input: &AcceptPortfolioShareInput
) -> Result<AcceptPortfolioShareOutput, AcceptPortfolioShareError>

Accepts an offer to share a portfolio.

fn associate_principal_with_portfolio(
&self,
input: &AssociatePrincipalWithPortfolioInput
) -> Result<AssociatePrincipalWithPortfolioOutput, AssociatePrincipalWithPortfolioError>

Associates the specified principal ARN with the specified portfolio.

fn associate_product_with_portfolio(
&self,
input: &AssociateProductWithPortfolioInput
) -> Result<AssociateProductWithPortfolioOutput, AssociateProductWithPortfolioError>

Associates a product with a portfolio.

fn create_constraint(
&self,
input: &CreateConstraintInput
) -> Result<CreateConstraintOutput, CreateConstraintError>

Creates a new constraint. For more information, see Using Constraints.

fn create_portfolio(
&self,
input: &CreatePortfolioInput
) -> Result<CreatePortfolioOutput, CreatePortfolioError>

Creates a new portfolio.

fn create_portfolio_share(
&self,
input: &CreatePortfolioShareInput
) -> Result<CreatePortfolioShareOutput, CreatePortfolioShareError>

Creates a new portfolio share.

fn create_product(
&self,
input: &CreateProductInput
) -> Result<CreateProductOutput, CreateProductError>

Creates a new product.

fn create_provisioning_artifact(
&self,
input: &CreateProvisioningArtifactInput
) -> Result<CreateProvisioningArtifactOutput, CreateProvisioningArtifactError>

Create a new provisioning artifact for the specified product. This operation does not work with a product that has been shared with you.

See the bottom of this topic for an example JSON request.

fn delete_constraint(
&self,
input: &DeleteConstraintInput
) -> Result<DeleteConstraintOutput, DeleteConstraintError>

Deletes the specified constraint.

fn delete_portfolio(
&self,
input: &DeletePortfolioInput
) -> Result<DeletePortfolioOutput, DeletePortfolioError>

Deletes the specified portfolio. This operation does not work with a portfolio that has been shared with you or if it has products, users, constraints, or shared accounts associated with it.

fn delete_portfolio_share(
&self,
input: &DeletePortfolioShareInput
) -> Result<DeletePortfolioShareOutput, DeletePortfolioShareError>

Deletes the specified portfolio share.

fn delete_product(
&self,
input: &DeleteProductInput
) -> Result<DeleteProductOutput, DeleteProductError>

Deletes the specified product. This operation does not work with a product that has been shared with you or is associated with a portfolio.

fn delete_provisioning_artifact(
&self,
input: &DeleteProvisioningArtifactInput
) -> Result<DeleteProvisioningArtifactOutput, DeleteProvisioningArtifactError>

Deletes the specified provisioning artifact. This operation does not work on a provisioning artifact associated with a product that has been shared with you, or on the last provisioning artifact associated with a product (a product must have at least one provisioning artifact).

fn describe_constraint(
&self,
input: &DescribeConstraintInput
) -> Result<DescribeConstraintOutput, DescribeConstraintError>

Retrieves detailed information for a specified constraint.

fn describe_portfolio(
&self,
input: &DescribePortfolioInput
) -> Result<DescribePortfolioOutput, DescribePortfolioError>

Retrieves detailed information and any tags associated with the specified portfolio.

fn describe_product(
&self,
input: &DescribeProductInput
) -> Result<DescribeProductOutput, DescribeProductError>

Retrieves information about a specified product.

This operation is functionally identical to DescribeProductView except that it takes as input ProductId instead of ProductViewId.

fn describe_product_as_admin(
&self,
input: &DescribeProductAsAdminInput
) -> Result<DescribeProductAsAdminOutput, DescribeProductAsAdminError>

Retrieves information about a specified product, run with administrator access.

fn describe_product_view(
&self,
input: &DescribeProductViewInput
) -> Result<DescribeProductViewOutput, DescribeProductViewError>

Retrieves information about a specified product.

This operation is functionally identical to DescribeProduct except that it takes as input ProductViewId instead of ProductId.

fn describe_provisioned_product(
&self,
input: &DescribeProvisionedProductInput
) -> Result<DescribeProvisionedProductOutput, DescribeProvisionedProductError>

Retrieve detailed information about the provisioned product.

fn describe_provisioning_artifact(
&self,
input: &DescribeProvisioningArtifactInput
) -> Result<DescribeProvisioningArtifactOutput, DescribeProvisioningArtifactError>

Retrieves detailed information about the specified provisioning artifact.

fn describe_provisioning_parameters(
&self,
input: &DescribeProvisioningParametersInput
) -> Result<DescribeProvisioningParametersOutput, DescribeProvisioningParametersError>

Provides information about parameters required to provision a specified product in a specified manner. Use this operation to obtain the list of ProvisioningArtifactParameters parameters available to call the ProvisionProduct operation for the specified product.

fn describe_record(
&self,
input: &DescribeRecordInput
) -> Result<DescribeRecordOutput, DescribeRecordError>

Retrieves a paginated list of the full details of a specific request. Use this operation after calling a request operation (ProvisionProduct, TerminateProvisionedProduct, or UpdateProvisionedProduct).

fn disassociate_principal_from_portfolio(
&self,
input: &DisassociatePrincipalFromPortfolioInput
) -> Result<DisassociatePrincipalFromPortfolioOutput, DisassociatePrincipalFromPortfolioError>

Disassociates a previously associated principal ARN from a specified portfolio.

fn disassociate_product_from_portfolio(
&self,
input: &DisassociateProductFromPortfolioInput
) -> Result<DisassociateProductFromPortfolioOutput, DisassociateProductFromPortfolioError>

Disassociates the specified product from the specified portfolio.

fn list_accepted_portfolio_shares(
&self,
input: &ListAcceptedPortfolioSharesInput
) -> Result<ListAcceptedPortfolioSharesOutput, ListAcceptedPortfolioSharesError>

Lists details of all portfolios for which sharing was accepted by this account.

fn list_constraints_for_portfolio(
&self,
input: &ListConstraintsForPortfolioInput
) -> Result<ListConstraintsForPortfolioOutput, ListConstraintsForPortfolioError>

Retrieves detailed constraint information for the specified portfolio and product.

fn list_launch_paths(
&self,
input: &ListLaunchPathsInput
) -> Result<ListLaunchPathsOutput, ListLaunchPathsError>

Returns a paginated list of all paths to a specified product. A path is how the user has access to a specified product, and is necessary when provisioning a product. A path also determines the constraints put on the product.

fn list_portfolio_access(
&self,
input: &ListPortfolioAccessInput
) -> Result<ListPortfolioAccessOutput, ListPortfolioAccessError>

Lists the account IDs that have been authorized sharing of the specified portfolio.

fn list_portfolios(
&self,
input: &ListPortfoliosInput
) -> Result<ListPortfoliosOutput, ListPortfoliosError>

Lists all portfolios in the catalog.

fn list_portfolios_for_product(
&self,
input: &ListPortfoliosForProductInput
) -> Result<ListPortfoliosForProductOutput, ListPortfoliosForProductError>

Lists all portfolios that the specified product is associated with.

fn list_principals_for_portfolio(
&self,
input: &ListPrincipalsForPortfolioInput
) -> Result<ListPrincipalsForPortfolioOutput, ListPrincipalsForPortfolioError>

Lists all principal ARNs associated with the specified portfolio.

fn list_provisioning_artifacts(
&self,
input: &ListProvisioningArtifactsInput
) -> Result<ListProvisioningArtifactsOutput, ListProvisioningArtifactsError>

Lists all provisioning artifacts associated with the specified product.

fn list_record_history(
&self,
input: &ListRecordHistoryInput
) -> Result<ListRecordHistoryOutput, ListRecordHistoryError>

Returns a paginated list of all performed requests, in the form of RecordDetails objects that are filtered as specified.

fn provision_product(
&self,
input: &ProvisionProductInput
) -> Result<ProvisionProductOutput, ProvisionProductError>

Requests a provision of a specified product. A provisioned product is a resourced instance for a product. For example, provisioning a CloudFormation-template-backed product results in launching a CloudFormation stack and all the underlying resources that come with it.

You can check the status of this request using the DescribeRecord operation.

fn reject_portfolio_share(
&self,
input: &RejectPortfolioShareInput
) -> Result<RejectPortfolioShareOutput, RejectPortfolioShareError>

Rejects an offer to share a portfolio.

fn scan_provisioned_products(
&self,
input: &ScanProvisionedProductsInput
) -> Result<ScanProvisionedProductsOutput, ScanProvisionedProductsError>

Returns a paginated list of all the ProvisionedProduct objects that are currently available (not terminated).

fn search_products(
&self,
input: &SearchProductsInput
) -> Result<SearchProductsOutput, SearchProductsError>

Returns a paginated list all of the Products objects to which the caller has access.

The output of this operation can be used as input for other operations, such as DescribeProductView.

fn search_products_as_admin(
&self,
input: &SearchProductsAsAdminInput
) -> Result<SearchProductsAsAdminOutput, SearchProductsAsAdminError>

Retrieves summary and status information about all products created within the caller's account. If a portfolio ID is provided, this operation retrieves information for only those products that are associated with the specified portfolio.

fn terminate_provisioned_product(
&self,
input: &TerminateProvisionedProductInput
) -> Result<TerminateProvisionedProductOutput, TerminateProvisionedProductError>

Requests termination of an existing ProvisionedProduct object. If there are Tags associated with the object, they are terminated when the ProvisionedProduct object is terminated.

This operation does not delete any records associated with the ProvisionedProduct object.

You can check the status of this request using the DescribeRecord operation.

fn update_constraint(
&self,
input: &UpdateConstraintInput
) -> Result<UpdateConstraintOutput, UpdateConstraintError>

Updates an existing constraint.

fn update_portfolio(
&self,
input: &UpdatePortfolioInput
) -> Result<UpdatePortfolioOutput, UpdatePortfolioError>

Updates the specified portfolio's details. This operation does not work with a product that has been shared with you.

fn update_product(
&self,
input: &UpdateProductInput
) -> Result<UpdateProductOutput, UpdateProductError>

Updates an existing product.

fn update_provisioned_product(
&self,
input: &UpdateProvisionedProductInput
) -> Result<UpdateProvisionedProductOutput, UpdateProvisionedProductError>

Requests updates to the configuration of an existing ProvisionedProduct object. If there are tags associated with the object, they cannot be updated or added with this operation. Depending on the specific updates requested, this operation may update with no interruption, with some interruption, or replace the ProvisionedProduct object entirely.

You can check the status of this request using the DescribeRecord operation.

fn update_provisioning_artifact(
&self,
input: &UpdateProvisioningArtifactInput
) -> Result<UpdateProvisioningArtifactOutput, UpdateProvisioningArtifactError>

Updates an existing provisioning artifact's information. This operation does not work on a provisioning artifact associated with a product that has been shared with you.