Crate rusoto_servicecatalog

AWS Service Catalog

If you're using the service, you're probably looking for ServiceCatalogClient and ServiceCatalog.

Structs

AcceptPortfolioShareInput

AcceptPortfolioShareOutput

AccessLevelFilter
The access level to limit results.

AssociatePrincipalWithPortfolioInput

AssociatePrincipalWithPortfolioOutput

AssociateProductWithPortfolioInput

AssociateProductWithPortfolioOutput

ConstraintDetail
Detailed constraint information.

ConstraintSummary
An administrator-specified constraint to apply when provisioning a product.

CreateConstraintInput

CreateConstraintOutput

CreatePortfolioInput

CreatePortfolioOutput

CreatePortfolioShareInput

CreatePortfolioShareOutput

CreateProductInput

CreateProductOutput

CreateProvisioningArtifactInput

CreateProvisioningArtifactOutput

DeleteConstraintInput

DeleteConstraintOutput

DeletePortfolioInput

DeletePortfolioOutput

DeletePortfolioShareInput

DeletePortfolioShareOutput

DeleteProductInput

DeleteProductOutput

DeleteProvisioningArtifactInput

DeleteProvisioningArtifactOutput

DescribeConstraintInput

DescribeConstraintOutput

DescribePortfolioInput

DescribePortfolioOutput

DescribeProductAsAdminInput

DescribeProductAsAdminOutput

DescribeProductInput

DescribeProductOutput

DescribeProductViewInput

DescribeProductViewOutput

DescribeProvisioningArtifactInput

DescribeProvisioningArtifactOutput

DescribeProvisioningParametersInput

DescribeProvisioningParametersOutput

DescribeRecordInput

DescribeRecordOutput

DisassociatePrincipalFromPortfolioInput

DisassociatePrincipalFromPortfolioOutput

DisassociateProductFromPortfolioInput

DisassociateProductFromPortfolioOutput

LaunchPathSummary
Summary information about a path for a user to have access to a specified product.

ListAcceptedPortfolioSharesInput

ListAcceptedPortfolioSharesOutput

ListConstraintsForPortfolioInput

ListConstraintsForPortfolioOutput

ListLaunchPathsInput

ListLaunchPathsOutput

ListPortfolioAccessInput

ListPortfolioAccessOutput

ListPortfoliosForProductInput

ListPortfoliosForProductOutput

ListPortfoliosInput

ListPortfoliosOutput

ListPrincipalsForPortfolioInput

ListPrincipalsForPortfolioOutput

ListProvisioningArtifactsInput

ListProvisioningArtifactsOutput

ListRecordHistoryInput

ListRecordHistoryOutput

ListRecordHistorySearchFilter
The search filter to limit results when listing request history records.

ParameterConstraints
The constraints that the administrator has put on the parameter.

PortfolioDetail
Detailed portfolio information.

Principal
A principal's ARN and type.

ProductViewAggregationValue
A single product view aggregation value/count pair, containing metadata about each product to which the calling user has access.

ProductViewDetail
Detailed product view information.

ProductViewSummary
The summary metadata about the specified product.

ProvisionProductInput

ProvisionProductOutput

ProvisionedProductDetail
Detailed information about a ProvisionedProduct object.

ProvisioningArtifact
Contains information indicating the ways in which a product can be provisioned.

ProvisioningArtifactDetail
Detailed provisioning artifact information.

ProvisioningArtifactParameter
A parameter used to successfully provision the product. This value includes a list of allowable values and additional metadata.

ProvisioningArtifactProperties
Provisioning artifact properties.

ProvisioningParameter
The arameter key/value pairs used to provision a product.

RecordDetail
The full details of a specific ProvisionedProduct object.

RecordError
The error code and description resulting from an operation.

RecordOutput
An output for the specified Product object created as the result of a request. For example, a CloudFormation-backed product that creates an S3 bucket would have an output for the S3 bucket URL.

RecordTag
A tag associated with the record, stored as a key-value pair.

RejectPortfolioShareInput

RejectPortfolioShareOutput

ScanProvisionedProductsInput

ScanProvisionedProductsOutput

SearchProductsAsAdminInput

SearchProductsAsAdminOutput

SearchProductsInput

SearchProductsOutput

ServiceCatalogClient
A client for the AWS Service Catalog API.

Tag
Key/value pairs to associate with this provisioning. These tags are entirely discretionary and are propagated to the resources created in the provisioning.

TerminateProvisionedProductInput

TerminateProvisionedProductOutput

UpdateConstraintInput

UpdateConstraintOutput

UpdatePortfolioInput

UpdatePortfolioOutput

UpdateProductInput

UpdateProductOutput

UpdateProvisionedProductInput

UpdateProvisionedProductOutput

UpdateProvisioningArtifactInput

UpdateProvisioningArtifactOutput

UpdateProvisioningParameter
The parameter key/value pair used to update a ProvisionedProduct object. If UsePreviousValue is set to true, Value is ignored and the value for Key is kept as previously set (current value).

UsageInstruction
Additional information provided by the administrator.

Enums

AcceptPortfolioShareError
Errors returned by AcceptPortfolioShare

AssociatePrincipalWithPortfolioError
Errors returned by AssociatePrincipalWithPortfolio

AssociateProductWithPortfolioError
Errors returned by AssociateProductWithPortfolio

CreateConstraintError
Errors returned by CreateConstraint

CreatePortfolioError
Errors returned by CreatePortfolio

CreatePortfolioShareError
Errors returned by CreatePortfolioShare

CreateProductError
Errors returned by CreateProduct

CreateProvisioningArtifactError
Errors returned by CreateProvisioningArtifact

DeleteConstraintError
Errors returned by DeleteConstraint

DeletePortfolioError
Errors returned by DeletePortfolio

DeletePortfolioShareError
Errors returned by DeletePortfolioShare

DeleteProductError
Errors returned by DeleteProduct

DeleteProvisioningArtifactError
Errors returned by DeleteProvisioningArtifact

DescribeConstraintError
Errors returned by DescribeConstraint

DescribePortfolioError
Errors returned by DescribePortfolio

DescribeProductAsAdminError
Errors returned by DescribeProductAsAdmin

DescribeProductError
Errors returned by DescribeProduct

DescribeProductViewError
Errors returned by DescribeProductView

DescribeProvisioningArtifactError
Errors returned by DescribeProvisioningArtifact

DescribeProvisioningParametersError
Errors returned by DescribeProvisioningParameters

DescribeRecordError
Errors returned by DescribeRecord

DisassociatePrincipalFromPortfolioError
Errors returned by DisassociatePrincipalFromPortfolio

DisassociateProductFromPortfolioError
Errors returned by DisassociateProductFromPortfolio

ListAcceptedPortfolioSharesError
Errors returned by ListAcceptedPortfolioShares

ListConstraintsForPortfolioError
Errors returned by ListConstraintsForPortfolio

ListLaunchPathsError
Errors returned by ListLaunchPaths

ListPortfolioAccessError
Errors returned by ListPortfolioAccess

ListPortfoliosError
Errors returned by ListPortfolios

ListPortfoliosForProductError
Errors returned by ListPortfoliosForProduct

ListPrincipalsForPortfolioError
Errors returned by ListPrincipalsForPortfolio

ListProvisioningArtifactsError
Errors returned by ListProvisioningArtifacts

ListRecordHistoryError
Errors returned by ListRecordHistory

ProvisionProductError
Errors returned by ProvisionProduct

RejectPortfolioShareError
Errors returned by RejectPortfolioShare

ScanProvisionedProductsError
Errors returned by ScanProvisionedProducts

SearchProductsAsAdminError
Errors returned by SearchProductsAsAdmin

SearchProductsError
Errors returned by SearchProducts

TerminateProvisionedProductError
Errors returned by TerminateProvisionedProduct

UpdateConstraintError
Errors returned by UpdateConstraint

UpdatePortfolioError
Errors returned by UpdatePortfolio

UpdateProductError
Errors returned by UpdateProduct

UpdateProvisionedProductError
Errors returned by UpdateProvisionedProduct

UpdateProvisioningArtifactError
Errors returned by UpdateProvisioningArtifact

Traits

ServiceCatalog
Trait representing the capabilities of the AWS Service Catalog API. AWS Service Catalog clients implement this trait.

Type Definitions

AcceptLanguage

AccessLevelFilterKey

AccessLevelFilterValue

AccountId

AccountIds

AddTags

AllowedValue

AllowedValues

ApproximateCount

AttributeValue

ConstraintDescription

ConstraintDetails

ConstraintParameters

ConstraintSummaries

ConstraintType

CreatedTime

CreationTime

DefaultValue

Description

ErrorCode

ErrorDescription

HasDefaultPath

Id

IdempotencyToken

IgnoreErrors

InstructionType

InstructionValue

LastRequestId

LaunchPathSummaries

NoEcho

NotificationArn

NotificationArns

OutputKey

OutputValue

PageSize

PageToken

ParameterKey

ParameterType

ParameterValue

PortfolioDescription

PortfolioDetails

PortfolioDisplayName

PortfolioName

PrincipalARN

PrincipalType

Principals

ProductSource

ProductType

ProductViewAggregationType

ProductViewAggregationValues

ProductViewAggregations

ProductViewDetails

ProductViewDistributor

ProductViewFilterBy

ProductViewFilterValue

ProductViewFilterValues

ProductViewFilters

ProductViewName

ProductViewOwner

ProductViewShortDescription

ProductViewSortBy

ProductViewSummaries

ProviderName

ProvisionedProductDetails

ProvisionedProductId

ProvisionedProductName

ProvisionedProductNameOrArn

ProvisionedProductStatusMessage

ProvisionedProductType

ProvisioningArtifactCreatedTime

ProvisioningArtifactDescription

ProvisioningArtifactDetails

ProvisioningArtifactInfo

ProvisioningArtifactInfoKey

ProvisioningArtifactInfoValue

ProvisioningArtifactName

ProvisioningArtifactParameters

ProvisioningArtifactType

ProvisioningArtifacts

ProvisioningParameters

RecordDetails

RecordErrors

RecordOutputs

RecordStatus

RecordTagKey

RecordTagValue

RecordTags

RecordType

ResourceARN

SearchFilterKey

SearchFilterValue

SortOrder

Status

SupportDescription

SupportEmail

SupportUrl

TagKey

TagKeys

TagValue

Tags

UpdateProvisioningParameters

UpdatedTime

UsageInstructions

UsePreviousValue