Crate rusoto_servicecatalog

AWS Service Catalog
AWS Service Catalog enables organizations to create and manage catalogs of IT services that are approved for use on AWS. To get the most out of this documentation, you should be familiar with the terminology discussed in AWS Service Catalog Concepts.

If you're using the service, you're probably looking for ServiceCatalogClient and ServiceCatalog.

Structs

AcceptPortfolioShareInput

AcceptPortfolioShareOutput

AccessLevelFilter
The access level to use to filter results.

AssociatePrincipalWithPortfolioInput

AssociatePrincipalWithPortfolioOutput

AssociateProductWithPortfolioInput

AssociateProductWithPortfolioOutput

AssociateTagOptionWithResourceInput

AssociateTagOptionWithResourceOutput

CloudWatchDashboard
Information about a CloudWatch dashboard.

ConstraintDetail
Information about a constraint.

ConstraintSummary
Summary information about a constraint.

CopyProductInput

CopyProductOutput

CreateConstraintInput

CreateConstraintOutput

CreatePortfolioInput

CreatePortfolioOutput

CreatePortfolioShareInput

CreatePortfolioShareOutput

CreateProductInput

CreateProductOutput

CreateProvisionedProductPlanInput

CreateProvisionedProductPlanOutput

CreateProvisioningArtifactInput

CreateProvisioningArtifactOutput

CreateTagOptionInput

CreateTagOptionOutput

DeleteConstraintInput

DeleteConstraintOutput

DeletePortfolioInput

DeletePortfolioOutput

DeletePortfolioShareInput

DeletePortfolioShareOutput

DeleteProductInput

DeleteProductOutput

DeleteProvisionedProductPlanInput

DeleteProvisionedProductPlanOutput

DeleteProvisioningArtifactInput

DeleteProvisioningArtifactOutput

DeleteTagOptionInput

DeleteTagOptionOutput

DescribeConstraintInput

DescribeConstraintOutput

DescribeCopyProductStatusInput

DescribeCopyProductStatusOutput

DescribePortfolioInput

DescribePortfolioOutput

DescribeProductAsAdminInput

DescribeProductAsAdminOutput

DescribeProductInput

DescribeProductOutput

DescribeProductViewInput

DescribeProductViewOutput

DescribeProvisionedProductInput

DescribeProvisionedProductOutput

DescribeProvisionedProductPlanInput

DescribeProvisionedProductPlanOutput

DescribeProvisioningArtifactInput

DescribeProvisioningArtifactOutput

DescribeProvisioningParametersInput

DescribeProvisioningParametersOutput

DescribeRecordInput

DescribeRecordOutput

DescribeTagOptionInput

DescribeTagOptionOutput

DisassociatePrincipalFromPortfolioInput

DisassociatePrincipalFromPortfolioOutput

DisassociateProductFromPortfolioInput

DisassociateProductFromPortfolioOutput

DisassociateTagOptionFromResourceInput

DisassociateTagOptionFromResourceOutput

ExecuteProvisionedProductPlanInput

ExecuteProvisionedProductPlanOutput

LaunchPathSummary
Summary information about a product path for a user.

ListAcceptedPortfolioSharesInput

ListAcceptedPortfolioSharesOutput

ListConstraintsForPortfolioInput

ListConstraintsForPortfolioOutput

ListLaunchPathsInput

ListLaunchPathsOutput

ListPortfolioAccessInput

ListPortfolioAccessOutput

ListPortfoliosForProductInput

ListPortfoliosForProductOutput

ListPortfoliosInput

ListPortfoliosOutput

ListPrincipalsForPortfolioInput

ListPrincipalsForPortfolioOutput

ListProvisionedProductPlansInput

ListProvisionedProductPlansOutput

ListProvisioningArtifactsInput

ListProvisioningArtifactsOutput

ListRecordHistoryInput

ListRecordHistoryOutput

ListRecordHistorySearchFilter
The search filter to use when listing history records.

ListResourcesForTagOptionInput

ListResourcesForTagOptionOutput

ListTagOptionsFilters
Filters to use when listing TagOptions.

ListTagOptionsInput

ListTagOptionsOutput

ParameterConstraints
The constraints that the administrator has put on the parameter.

PortfolioDetail
Information about a portfolio.

Principal
Information about a principal.

ProductViewAggregationValue
A single product view aggregation value/count pair, containing metadata about each product to which the calling user has access.

ProductViewDetail
Information about a product view.

ProductViewSummary
Summary information about a product view.

ProvisionProductInput

ProvisionProductOutput

ProvisionedProductAttribute
Information about a provisioned product.

ProvisionedProductDetail
Information about a provisioned product.

ProvisionedProductPlanDetails
Information about a plan.

ProvisionedProductPlanSummary
Summary information about a plan.

ProvisioningArtifact
Information about a provisioning artifact. A provisioning artifact is also known as a product version.

ProvisioningArtifactDetail
Information about a provisioning artifact (also known as a version) for a product.

ProvisioningArtifactParameter
Information about a parameter used to provision a product.

ProvisioningArtifactProperties
Information about a provisioning artifact (also known as a version) for a product.

ProvisioningArtifactSummary
Summary information about a provisioning artifact (also known as a version) for a product.

ProvisioningParameter
Information about a parameter used to provision a product.

RecordDetail
Information about a request operation.

RecordError
The error code and description resulting from an operation.

RecordOutput
The output for the product created as the result of a request. For example, the output for a CloudFormation-backed product that creates an S3 bucket would include the S3 bucket URL.

RecordTag
Information about a tag, which is a key-value pair.

RejectPortfolioShareInput

RejectPortfolioShareOutput

ResourceChange
Information about a resource change that will occur when a plan is executed.

ResourceChangeDetail
Information about a change to a resource attribute.

ResourceDetail
Information about a resource.

ResourceTargetDefinition
Information about a change to a resource attribute.

ScanProvisionedProductsInput

ScanProvisionedProductsOutput

SearchProductsAsAdminInput

SearchProductsAsAdminOutput

SearchProductsInput

SearchProductsOutput

SearchProvisionedProductsInput

SearchProvisionedProductsOutput

ServiceCatalogClient
A client for the AWS Service Catalog API.

Tag
Information about a tag. A tag is a key-value pair. Tags are propagated to the resources created when provisioning a product.

TagOptionDetail
Information about a TagOption.

TagOptionSummary
Summary information about a TagOption.

TerminateProvisionedProductInput

TerminateProvisionedProductOutput

UpdateConstraintInput

UpdateConstraintOutput

UpdatePortfolioInput

UpdatePortfolioOutput

UpdateProductInput

UpdateProductOutput

UpdateProvisionedProductInput

UpdateProvisionedProductOutput

UpdateProvisioningArtifactInput

UpdateProvisioningArtifactOutput

UpdateProvisioningParameter
The parameter key-value pair used to update a provisioned product.

UpdateTagOptionInput

UpdateTagOptionOutput

UsageInstruction
Additional information provided by the administrator.

Enums

AcceptPortfolioShareError
Errors returned by AcceptPortfolioShare

AssociatePrincipalWithPortfolioError
Errors returned by AssociatePrincipalWithPortfolio

AssociateProductWithPortfolioError
Errors returned by AssociateProductWithPortfolio

AssociateTagOptionWithResourceError
Errors returned by AssociateTagOptionWithResource

CopyProductError
Errors returned by CopyProduct

CreateConstraintError
Errors returned by CreateConstraint

CreatePortfolioError
Errors returned by CreatePortfolio

CreatePortfolioShareError
Errors returned by CreatePortfolioShare

CreateProductError
Errors returned by CreateProduct

CreateProvisionedProductPlanError
Errors returned by CreateProvisionedProductPlan

CreateProvisioningArtifactError
Errors returned by CreateProvisioningArtifact

CreateTagOptionError
Errors returned by CreateTagOption

DeleteConstraintError
Errors returned by DeleteConstraint

DeletePortfolioError
Errors returned by DeletePortfolio

DeletePortfolioShareError
Errors returned by DeletePortfolioShare

DeleteProductError
Errors returned by DeleteProduct

DeleteProvisionedProductPlanError
Errors returned by DeleteProvisionedProductPlan

DeleteProvisioningArtifactError
Errors returned by DeleteProvisioningArtifact

DeleteTagOptionError
Errors returned by DeleteTagOption

DescribeConstraintError
Errors returned by DescribeConstraint

DescribeCopyProductStatusError
Errors returned by DescribeCopyProductStatus

DescribePortfolioError
Errors returned by DescribePortfolio

DescribeProductAsAdminError
Errors returned by DescribeProductAsAdmin

DescribeProductError
Errors returned by DescribeProduct

DescribeProductViewError
Errors returned by DescribeProductView

DescribeProvisionedProductError
Errors returned by DescribeProvisionedProduct

DescribeProvisionedProductPlanError
Errors returned by DescribeProvisionedProductPlan

DescribeProvisioningArtifactError
Errors returned by DescribeProvisioningArtifact

DescribeProvisioningParametersError
Errors returned by DescribeProvisioningParameters

DescribeRecordError
Errors returned by DescribeRecord

DescribeTagOptionError
Errors returned by DescribeTagOption

DisassociatePrincipalFromPortfolioError
Errors returned by DisassociatePrincipalFromPortfolio

DisassociateProductFromPortfolioError
Errors returned by DisassociateProductFromPortfolio

DisassociateTagOptionFromResourceError
Errors returned by DisassociateTagOptionFromResource

ExecuteProvisionedProductPlanError
Errors returned by ExecuteProvisionedProductPlan

ListAcceptedPortfolioSharesError
Errors returned by ListAcceptedPortfolioShares

ListConstraintsForPortfolioError
Errors returned by ListConstraintsForPortfolio

ListLaunchPathsError
Errors returned by ListLaunchPaths

ListPortfolioAccessError
Errors returned by ListPortfolioAccess

ListPortfoliosError
Errors returned by ListPortfolios

ListPortfoliosForProductError
Errors returned by ListPortfoliosForProduct

ListPrincipalsForPortfolioError
Errors returned by ListPrincipalsForPortfolio

ListProvisionedProductPlansError
Errors returned by ListProvisionedProductPlans

ListProvisioningArtifactsError
Errors returned by ListProvisioningArtifacts

ListRecordHistoryError
Errors returned by ListRecordHistory

ListResourcesForTagOptionError
Errors returned by ListResourcesForTagOption

ListTagOptionsError
Errors returned by ListTagOptions

ProvisionProductError
Errors returned by ProvisionProduct

RejectPortfolioShareError
Errors returned by RejectPortfolioShare

ScanProvisionedProductsError
Errors returned by ScanProvisionedProducts

SearchProductsAsAdminError
Errors returned by SearchProductsAsAdmin

SearchProductsError
Errors returned by SearchProducts

SearchProvisionedProductsError
Errors returned by SearchProvisionedProducts

TerminateProvisionedProductError
Errors returned by TerminateProvisionedProduct

UpdateConstraintError
Errors returned by UpdateConstraint

UpdatePortfolioError
Errors returned by UpdatePortfolio

UpdateProductError
Errors returned by UpdateProduct

UpdateProvisionedProductError
Errors returned by UpdateProvisionedProduct

UpdateProvisioningArtifactError
Errors returned by UpdateProvisioningArtifact

UpdateTagOptionError
Errors returned by UpdateTagOption

Traits

ServiceCatalog
Trait representing the capabilities of the AWS Service Catalog API. AWS Service Catalog clients implement this trait.