Trait rusoto_servicecatalog::ServiceCatalog

pub trait ServiceCatalog {
    fn accept_portfolio_share(
        &self, 
        input: AcceptPortfolioShareInput
    ) -> RusotoFuture<AcceptPortfolioShareOutput, AcceptPortfolioShareError>;

    fn associate_principal_with_portfolio(
        &self, 
        input: AssociatePrincipalWithPortfolioInput
    ) -> RusotoFuture<AssociatePrincipalWithPortfolioOutput, AssociatePrincipalWithPortfolioError>;

    fn associate_product_with_portfolio(
        &self, 
        input: AssociateProductWithPortfolioInput
    ) -> RusotoFuture<AssociateProductWithPortfolioOutput, AssociateProductWithPortfolioError>;

    fn associate_tag_option_with_resource(
        &self, 
        input: AssociateTagOptionWithResourceInput
    ) -> RusotoFuture<AssociateTagOptionWithResourceOutput, AssociateTagOptionWithResourceError>;

    fn copy_product(
        &self, 
        input: CopyProductInput
    ) -> RusotoFuture<CopyProductOutput, CopyProductError>;

    fn create_constraint(
        &self, 
        input: CreateConstraintInput
    ) -> RusotoFuture<CreateConstraintOutput, CreateConstraintError>;

    fn create_portfolio(
        &self, 
        input: CreatePortfolioInput
    ) -> RusotoFuture<CreatePortfolioOutput, CreatePortfolioError>;

    fn create_portfolio_share(
        &self, 
        input: CreatePortfolioShareInput
    ) -> RusotoFuture<CreatePortfolioShareOutput, CreatePortfolioShareError>;

    fn create_product(
        &self, 
        input: CreateProductInput
    ) -> RusotoFuture<CreateProductOutput, CreateProductError>;

    fn create_provisioned_product_plan(
        &self, 
        input: CreateProvisionedProductPlanInput
    ) -> RusotoFuture<CreateProvisionedProductPlanOutput, CreateProvisionedProductPlanError>;

    fn create_provisioning_artifact(
        &self, 
        input: CreateProvisioningArtifactInput
    ) -> RusotoFuture<CreateProvisioningArtifactOutput, CreateProvisioningArtifactError>;

    fn create_tag_option(
        &self, 
        input: CreateTagOptionInput
    ) -> RusotoFuture<CreateTagOptionOutput, CreateTagOptionError>;

    fn delete_constraint(
        &self, 
        input: DeleteConstraintInput
    ) -> RusotoFuture<DeleteConstraintOutput, DeleteConstraintError>;

    fn delete_portfolio(
        &self, 
        input: DeletePortfolioInput
    ) -> RusotoFuture<DeletePortfolioOutput, DeletePortfolioError>;

    fn delete_portfolio_share(
        &self, 
        input: DeletePortfolioShareInput
    ) -> RusotoFuture<DeletePortfolioShareOutput, DeletePortfolioShareError>;

    fn delete_product(
        &self, 
        input: DeleteProductInput
    ) -> RusotoFuture<DeleteProductOutput, DeleteProductError>;

    fn delete_provisioned_product_plan(
        &self, 
        input: DeleteProvisionedProductPlanInput
    ) -> RusotoFuture<DeleteProvisionedProductPlanOutput, DeleteProvisionedProductPlanError>;

    fn delete_provisioning_artifact(
        &self, 
        input: DeleteProvisioningArtifactInput
    ) -> RusotoFuture<DeleteProvisioningArtifactOutput, DeleteProvisioningArtifactError>;

    fn delete_tag_option(
        &self, 
        input: DeleteTagOptionInput
    ) -> RusotoFuture<DeleteTagOptionOutput, DeleteTagOptionError>;

    fn describe_constraint(
        &self, 
        input: DescribeConstraintInput
    ) -> RusotoFuture<DescribeConstraintOutput, DescribeConstraintError>;

    fn describe_copy_product_status(
        &self, 
        input: DescribeCopyProductStatusInput
    ) -> RusotoFuture<DescribeCopyProductStatusOutput, DescribeCopyProductStatusError>;

    fn describe_portfolio(
        &self, 
        input: DescribePortfolioInput
    ) -> RusotoFuture<DescribePortfolioOutput, DescribePortfolioError>;

    fn describe_product(
        &self, 
        input: DescribeProductInput
    ) -> RusotoFuture<DescribeProductOutput, DescribeProductError>;

    fn describe_product_as_admin(
        &self, 
        input: DescribeProductAsAdminInput
    ) -> RusotoFuture<DescribeProductAsAdminOutput, DescribeProductAsAdminError>;

    fn describe_product_view(
        &self, 
        input: DescribeProductViewInput
    ) -> RusotoFuture<DescribeProductViewOutput, DescribeProductViewError>;

    fn describe_provisioned_product(
        &self, 
        input: DescribeProvisionedProductInput
    ) -> RusotoFuture<DescribeProvisionedProductOutput, DescribeProvisionedProductError>;

    fn describe_provisioned_product_plan(
        &self, 
        input: DescribeProvisionedProductPlanInput
    ) -> RusotoFuture<DescribeProvisionedProductPlanOutput, DescribeProvisionedProductPlanError>;

    fn describe_provisioning_artifact(
        &self, 
        input: DescribeProvisioningArtifactInput
    ) -> RusotoFuture<DescribeProvisioningArtifactOutput, DescribeProvisioningArtifactError>;

    fn describe_provisioning_parameters(
        &self, 
        input: DescribeProvisioningParametersInput
    ) -> RusotoFuture<DescribeProvisioningParametersOutput, DescribeProvisioningParametersError>;

    fn describe_record(
        &self, 
        input: DescribeRecordInput
    ) -> RusotoFuture<DescribeRecordOutput, DescribeRecordError>;

    fn describe_tag_option(
        &self, 
        input: DescribeTagOptionInput
    ) -> RusotoFuture<DescribeTagOptionOutput, DescribeTagOptionError>;

    fn disassociate_principal_from_portfolio(
        &self, 
        input: DisassociatePrincipalFromPortfolioInput
    ) -> RusotoFuture<DisassociatePrincipalFromPortfolioOutput, DisassociatePrincipalFromPortfolioError>;

    fn disassociate_product_from_portfolio(
        &self, 
        input: DisassociateProductFromPortfolioInput
    ) -> RusotoFuture<DisassociateProductFromPortfolioOutput, DisassociateProductFromPortfolioError>;

    fn disassociate_tag_option_from_resource(
        &self, 
        input: DisassociateTagOptionFromResourceInput
    ) -> RusotoFuture<DisassociateTagOptionFromResourceOutput, DisassociateTagOptionFromResourceError>;

    fn execute_provisioned_product_plan(
        &self, 
        input: ExecuteProvisionedProductPlanInput
    ) -> RusotoFuture<ExecuteProvisionedProductPlanOutput, ExecuteProvisionedProductPlanError>;

    fn list_accepted_portfolio_shares(
        &self, 
        input: ListAcceptedPortfolioSharesInput
    ) -> RusotoFuture<ListAcceptedPortfolioSharesOutput, ListAcceptedPortfolioSharesError>;

    fn list_constraints_for_portfolio(
        &self, 
        input: ListConstraintsForPortfolioInput
    ) -> RusotoFuture<ListConstraintsForPortfolioOutput, ListConstraintsForPortfolioError>;

    fn list_launch_paths(
        &self, 
        input: ListLaunchPathsInput
    ) -> RusotoFuture<ListLaunchPathsOutput, ListLaunchPathsError>;

    fn list_portfolio_access(
        &self, 
        input: ListPortfolioAccessInput
    ) -> RusotoFuture<ListPortfolioAccessOutput, ListPortfolioAccessError>;

    fn list_portfolios(
        &self, 
        input: ListPortfoliosInput
    ) -> RusotoFuture<ListPortfoliosOutput, ListPortfoliosError>;

    fn list_portfolios_for_product(
        &self, 
        input: ListPortfoliosForProductInput
    ) -> RusotoFuture<ListPortfoliosForProductOutput, ListPortfoliosForProductError>;

    fn list_principals_for_portfolio(
        &self, 
        input: ListPrincipalsForPortfolioInput
    ) -> RusotoFuture<ListPrincipalsForPortfolioOutput, ListPrincipalsForPortfolioError>;

    fn list_provisioned_product_plans(
        &self, 
        input: ListProvisionedProductPlansInput
    ) -> RusotoFuture<ListProvisionedProductPlansOutput, ListProvisionedProductPlansError>;

    fn list_provisioning_artifacts(
        &self, 
        input: ListProvisioningArtifactsInput
    ) -> RusotoFuture<ListProvisioningArtifactsOutput, ListProvisioningArtifactsError>;

    fn list_record_history(
        &self, 
        input: ListRecordHistoryInput
    ) -> RusotoFuture<ListRecordHistoryOutput, ListRecordHistoryError>;

    fn list_resources_for_tag_option(
        &self, 
        input: ListResourcesForTagOptionInput
    ) -> RusotoFuture<ListResourcesForTagOptionOutput, ListResourcesForTagOptionError>;

    fn list_tag_options(
        &self, 
        input: ListTagOptionsInput
    ) -> RusotoFuture<ListTagOptionsOutput, ListTagOptionsError>;

    fn provision_product(
        &self, 
        input: ProvisionProductInput
    ) -> RusotoFuture<ProvisionProductOutput, ProvisionProductError>;

    fn reject_portfolio_share(
        &self, 
        input: RejectPortfolioShareInput
    ) -> RusotoFuture<RejectPortfolioShareOutput, RejectPortfolioShareError>;

    fn scan_provisioned_products(
        &self, 
        input: ScanProvisionedProductsInput
    ) -> RusotoFuture<ScanProvisionedProductsOutput, ScanProvisionedProductsError>;

    fn search_products(
        &self, 
        input: SearchProductsInput
    ) -> RusotoFuture<SearchProductsOutput, SearchProductsError>;

    fn search_products_as_admin(
        &self, 
        input: SearchProductsAsAdminInput
    ) -> RusotoFuture<SearchProductsAsAdminOutput, SearchProductsAsAdminError>;

    fn search_provisioned_products(
        &self, 
        input: SearchProvisionedProductsInput
    ) -> RusotoFuture<SearchProvisionedProductsOutput, SearchProvisionedProductsError>;

    fn terminate_provisioned_product(
        &self, 
        input: TerminateProvisionedProductInput
    ) -> RusotoFuture<TerminateProvisionedProductOutput, TerminateProvisionedProductError>;

    fn update_constraint(
        &self, 
        input: UpdateConstraintInput
    ) -> RusotoFuture<UpdateConstraintOutput, UpdateConstraintError>;

    fn update_portfolio(
        &self, 
        input: UpdatePortfolioInput
    ) -> RusotoFuture<UpdatePortfolioOutput, UpdatePortfolioError>;

    fn update_product(
        &self, 
        input: UpdateProductInput
    ) -> RusotoFuture<UpdateProductOutput, UpdateProductError>;

    fn update_provisioned_product(
        &self, 
        input: UpdateProvisionedProductInput
    ) -> RusotoFuture<UpdateProvisionedProductOutput, UpdateProvisionedProductError>;

    fn update_provisioning_artifact(
        &self, 
        input: UpdateProvisioningArtifactInput
    ) -> RusotoFuture<UpdateProvisioningArtifactOutput, UpdateProvisioningArtifactError>;

    fn update_tag_option(
        &self, 
        input: UpdateTagOptionInput
    ) -> RusotoFuture<UpdateTagOptionOutput, UpdateTagOptionError>;
}

Trait representing the capabilities of the AWS Service Catalog API. AWS Service Catalog clients implement this trait.

Required Methods

fn accept_portfolio_share(
&self,
input: AcceptPortfolioShareInput
) -> RusotoFuture<AcceptPortfolioShareOutput, AcceptPortfolioShareError>

Accepts an offer to share the specified portfolio.

fn associate_principal_with_portfolio(
&self,
input: AssociatePrincipalWithPortfolioInput
) -> RusotoFuture<AssociatePrincipalWithPortfolioOutput, AssociatePrincipalWithPortfolioError>

Associates the specified principal ARN with the specified portfolio.

fn associate_product_with_portfolio(
&self,
input: AssociateProductWithPortfolioInput
) -> RusotoFuture<AssociateProductWithPortfolioOutput, AssociateProductWithPortfolioError>

Associates the specified product with the specified portfolio.

fn associate_tag_option_with_resource(
&self,
input: AssociateTagOptionWithResourceInput
) -> RusotoFuture<AssociateTagOptionWithResourceOutput, AssociateTagOptionWithResourceError>

Associate the specified TagOption with the specified portfolio or product.

fn copy_product(
&self,
input: CopyProductInput
) -> RusotoFuture<CopyProductOutput, CopyProductError>

Copies the specified source product to the specified target product or a new product.

You can copy a product to the same account or another account. You can copy a product to the same region or another region.

This operation is performed asynchronously. To track the progress of the operation, use DescribeCopyProductStatus.

fn create_constraint(
&self,
input: CreateConstraintInput
) -> RusotoFuture<CreateConstraintOutput, CreateConstraintError>

Creates a constraint.

fn create_portfolio(
&self,
input: CreatePortfolioInput
) -> RusotoFuture<CreatePortfolioOutput, CreatePortfolioError>

Creates a portfolio.

fn create_portfolio_share(
&self,
input: CreatePortfolioShareInput
) -> RusotoFuture<CreatePortfolioShareOutput, CreatePortfolioShareError>

Shares the specified portfolio with the specified account.

fn create_product(
&self,
input: CreateProductInput
) -> RusotoFuture<CreateProductOutput, CreateProductError>

Creates a product.

fn create_provisioned_product_plan(
&self,
input: CreateProvisionedProductPlanInput
) -> RusotoFuture<CreateProvisionedProductPlanOutput, CreateProvisionedProductPlanError>

Creates a plan. A plan includes the list of resources to be created (when provisioning a new product) or modified (when updating a provisioned product) when the plan is executed.

You can create one plan per provisioned product. To create a plan for an existing provisioned product, the product status must be AVAILBLE or TAINTED.

To view the resource changes in the change set, use DescribeProvisionedProductPlan. To create or modify the provisioned product, use ExecuteProvisionedProductPlan.

fn create_provisioning_artifact(
&self,
input: CreateProvisioningArtifactInput
) -> RusotoFuture<CreateProvisioningArtifactOutput, CreateProvisioningArtifactError>

Creates a provisioning artifact (also known as a version) for the specified product.

You cannot create a provisioning artifact for a product that was shared with you.

fn create_tag_option(
&self,
input: CreateTagOptionInput
) -> RusotoFuture<CreateTagOptionOutput, CreateTagOptionError>

Creates a TagOption.

fn delete_constraint(
&self,
input: DeleteConstraintInput
) -> RusotoFuture<DeleteConstraintOutput, DeleteConstraintError>

Deletes the specified constraint.

fn delete_portfolio(
&self,
input: DeletePortfolioInput
) -> RusotoFuture<DeletePortfolioOutput, DeletePortfolioError>

Deletes the specified portfolio.

You cannot delete a portfolio if it was shared with you or if it has associated products, users, constraints, or shared accounts.

fn delete_portfolio_share(
&self,
input: DeletePortfolioShareInput
) -> RusotoFuture<DeletePortfolioShareOutput, DeletePortfolioShareError>

Stops sharing the specified portfolio with the specified account.

fn delete_product(
&self,
input: DeleteProductInput
) -> RusotoFuture<DeleteProductOutput, DeleteProductError>

Deletes the specified product.

You cannot delete a product if it was shared with you or is associated with a portfolio.

fn delete_provisioned_product_plan(
&self,
input: DeleteProvisionedProductPlanInput
) -> RusotoFuture<DeleteProvisionedProductPlanOutput, DeleteProvisionedProductPlanError>

Deletes the specified plan.

fn delete_provisioning_artifact(
&self,
input: DeleteProvisioningArtifactInput
) -> RusotoFuture<DeleteProvisioningArtifactOutput, DeleteProvisioningArtifactError>

Deletes the specified provisioning artifact (also known as a version) for the specified product.

You cannot delete a provisioning artifact associated with a product that was shared with you. You cannot delete the last provisioning artifact for a product, because a product must have at least one provisioning artifact.

fn delete_tag_option(
&self,
input: DeleteTagOptionInput
) -> RusotoFuture<DeleteTagOptionOutput, DeleteTagOptionError>

Deletes the specified TagOption.

You cannot delete a TagOption if it is associated with a product or portfolio.

fn describe_constraint(
&self,
input: DescribeConstraintInput
) -> RusotoFuture<DescribeConstraintOutput, DescribeConstraintError>

Gets information about the specified constraint.

fn describe_copy_product_status(
&self,
input: DescribeCopyProductStatusInput
) -> RusotoFuture<DescribeCopyProductStatusOutput, DescribeCopyProductStatusError>

Gets the status of the specified copy product operation.

fn describe_portfolio(
&self,
input: DescribePortfolioInput
) -> RusotoFuture<DescribePortfolioOutput, DescribePortfolioError>

Gets information about the specified portfolio.

fn describe_product(
&self,
input: DescribeProductInput
) -> RusotoFuture<DescribeProductOutput, DescribeProductError>

Gets information about the specified product.

fn describe_product_as_admin(
&self,
input: DescribeProductAsAdminInput
) -> RusotoFuture<DescribeProductAsAdminOutput, DescribeProductAsAdminError>

Gets information about the specified product. This operation is run with administrator access.

fn describe_product_view(
&self,
input: DescribeProductViewInput
) -> RusotoFuture<DescribeProductViewOutput, DescribeProductViewError>

Gets information about the specified product.

fn describe_provisioned_product(
&self,
input: DescribeProvisionedProductInput
) -> RusotoFuture<DescribeProvisionedProductOutput, DescribeProvisionedProductError>

Gets information about the specified provisioned product.

fn describe_provisioned_product_plan(
&self,
input: DescribeProvisionedProductPlanInput
) -> RusotoFuture<DescribeProvisionedProductPlanOutput, DescribeProvisionedProductPlanError>

Gets information about the resource changes for the specified plan.

fn describe_provisioning_artifact(
&self,
input: DescribeProvisioningArtifactInput
) -> RusotoFuture<DescribeProvisioningArtifactOutput, DescribeProvisioningArtifactError>

Gets information about the specified provisioning artifact (also known as a version) for the specified product.

fn describe_provisioning_parameters(
&self,
input: DescribeProvisioningParametersInput
) -> RusotoFuture<DescribeProvisioningParametersOutput, DescribeProvisioningParametersError>

Gets information about the configuration required to provision the specified product using the specified provisioning artifact.

If the output contains a TagOption key with an empty list of values, there is a TagOption conflict for that key. The end user cannot take action to fix the conflict, and launch is not blocked. In subsequent calls to ProvisionProduct, do not include conflicted TagOption keys as tags, or this causes the error "Parameter validation failed: Missing required parameter in Tags[N]:Value". Tag the provisioned product with the value sc-tagoption-conflict-portfolioId-productId.

fn describe_record(
&self,
input: DescribeRecordInput
) -> RusotoFuture<DescribeRecordOutput, DescribeRecordError>

Gets information about the specified request operation.

Use this operation after calling a request operation (for example, ProvisionProduct, TerminateProvisionedProduct, or UpdateProvisionedProduct).

fn describe_tag_option(
&self,
input: DescribeTagOptionInput
) -> RusotoFuture<DescribeTagOptionOutput, DescribeTagOptionError>

Gets information about the specified TagOption.

fn disassociate_principal_from_portfolio(
&self,
input: DisassociatePrincipalFromPortfolioInput
) -> RusotoFuture<DisassociatePrincipalFromPortfolioOutput, DisassociatePrincipalFromPortfolioError>

Disassociates a previously associated principal ARN from a specified portfolio.

fn disassociate_product_from_portfolio(
&self,
input: DisassociateProductFromPortfolioInput
) -> RusotoFuture<DisassociateProductFromPortfolioOutput, DisassociateProductFromPortfolioError>

Disassociates the specified product from the specified portfolio.

fn disassociate_tag_option_from_resource(
&self,
input: DisassociateTagOptionFromResourceInput
) -> RusotoFuture<DisassociateTagOptionFromResourceOutput, DisassociateTagOptionFromResourceError>

Disassociates the specified TagOption from the specified resource.

fn execute_provisioned_product_plan(
&self,
input: ExecuteProvisionedProductPlanInput
) -> RusotoFuture<ExecuteProvisionedProductPlanOutput, ExecuteProvisionedProductPlanError>

Provisions or modifies a product based on the resource changes for the specified plan.

fn list_accepted_portfolio_shares(
&self,
input: ListAcceptedPortfolioSharesInput
) -> RusotoFuture<ListAcceptedPortfolioSharesOutput, ListAcceptedPortfolioSharesError>

Lists all portfolios for which sharing was accepted by this account.

fn list_constraints_for_portfolio(
&self,
input: ListConstraintsForPortfolioInput
) -> RusotoFuture<ListConstraintsForPortfolioOutput, ListConstraintsForPortfolioError>

Lists the constraints for the specified portfolio and product.

fn list_launch_paths(
&self,
input: ListLaunchPathsInput
) -> RusotoFuture<ListLaunchPathsOutput, ListLaunchPathsError>

Lists the paths to the specified product. A path is how the user has access to a specified product, and is necessary when provisioning a product. A path also determines the constraints put on the product.

fn list_portfolio_access(
&self,
input: ListPortfolioAccessInput
) -> RusotoFuture<ListPortfolioAccessOutput, ListPortfolioAccessError>

Lists the account IDs that have access to the specified portfolio.

fn list_portfolios(
&self,
input: ListPortfoliosInput
) -> RusotoFuture<ListPortfoliosOutput, ListPortfoliosError>

Lists all portfolios in the catalog.

fn list_portfolios_for_product(
&self,
input: ListPortfoliosForProductInput
) -> RusotoFuture<ListPortfoliosForProductOutput, ListPortfoliosForProductError>

Lists all portfolios that the specified product is associated with.

fn list_principals_for_portfolio(
&self,
input: ListPrincipalsForPortfolioInput
) -> RusotoFuture<ListPrincipalsForPortfolioOutput, ListPrincipalsForPortfolioError>

Lists all principal ARNs associated with the specified portfolio.

fn list_provisioned_product_plans(
&self,
input: ListProvisionedProductPlansInput
) -> RusotoFuture<ListProvisionedProductPlansOutput, ListProvisionedProductPlansError>

Lists the plans for the specified provisioned product or all plans to which the user has access.

fn list_provisioning_artifacts(
&self,
input: ListProvisioningArtifactsInput
) -> RusotoFuture<ListProvisioningArtifactsOutput, ListProvisioningArtifactsError>

Lists all provisioning artifacts (also known as versions) for the specified product.

fn list_record_history(
&self,
input: ListRecordHistoryInput
) -> RusotoFuture<ListRecordHistoryOutput, ListRecordHistoryError>

Lists the specified requests or all performed requests.

fn list_resources_for_tag_option(
&self,
input: ListResourcesForTagOptionInput
) -> RusotoFuture<ListResourcesForTagOptionOutput, ListResourcesForTagOptionError>

Lists the resources associated with the specified TagOption.

fn list_tag_options(
&self,
input: ListTagOptionsInput
) -> RusotoFuture<ListTagOptionsOutput, ListTagOptionsError>

Lists the specified TagOptions or all TagOptions.

fn provision_product(
&self,
input: ProvisionProductInput
) -> RusotoFuture<ProvisionProductOutput, ProvisionProductError>

Provisions the specified product.

A provisioned product is a resourced instance of a product. For example, provisioning a product based on a CloudFormation template launches a CloudFormation stack and its underlying resources. You can check the status of this request using DescribeRecord.

If the request contains a tag key with an empty list of values, there is a tag conflict for that key. Do not include conflicted keys as tags, or this causes the error "Parameter validation failed: Missing required parameter in Tags[N]:Value".

fn reject_portfolio_share(
&self,
input: RejectPortfolioShareInput
) -> RusotoFuture<RejectPortfolioShareOutput, RejectPortfolioShareError>

Rejects an offer to share the specified portfolio.

fn scan_provisioned_products(
&self,
input: ScanProvisionedProductsInput
) -> RusotoFuture<ScanProvisionedProductsOutput, ScanProvisionedProductsError>

Lists the provisioned products that are available (not terminated).

To use additional filtering, see SearchProvisionedProducts.

fn search_products(
&self,
input: SearchProductsInput
) -> RusotoFuture<SearchProductsOutput, SearchProductsError>

Gets information about the products to which the caller has access.

fn search_products_as_admin(
&self,
input: SearchProductsAsAdminInput
) -> RusotoFuture<SearchProductsAsAdminOutput, SearchProductsAsAdminError>

Gets information about the products for the specified portfolio or all products.

fn search_provisioned_products(
&self,
input: SearchProvisionedProductsInput
) -> RusotoFuture<SearchProvisionedProductsOutput, SearchProvisionedProductsError>

Gets information about the provisioned products that meet the specified criteria.

fn terminate_provisioned_product(
&self,
input: TerminateProvisionedProductInput
) -> RusotoFuture<TerminateProvisionedProductOutput, TerminateProvisionedProductError>

Terminates the specified provisioned product.

This operation does not delete any records associated with the provisioned product.

You can check the status of this request using DescribeRecord.

fn update_constraint(
&self,
input: UpdateConstraintInput
) -> RusotoFuture<UpdateConstraintOutput, UpdateConstraintError>

Updates the specified constraint.

fn update_portfolio(
&self,
input: UpdatePortfolioInput
) -> RusotoFuture<UpdatePortfolioOutput, UpdatePortfolioError>

Updates the specified portfolio.

You cannot update a product that was shared with you.

fn update_product(
&self,
input: UpdateProductInput
) -> RusotoFuture<UpdateProductOutput, UpdateProductError>

Updates the specified product.

fn update_provisioned_product(
&self,
input: UpdateProvisionedProductInput
) -> RusotoFuture<UpdateProvisionedProductOutput, UpdateProvisionedProductError>

Requests updates to the configuration of the specified provisioned product.

If there are tags associated with the object, they cannot be updated or added. Depending on the specific updates requested, this operation can update with no interruption, with some interruption, or replace the provisioned product entirely.

You can check the status of this request using DescribeRecord.

fn update_provisioning_artifact(
&self,
input: UpdateProvisioningArtifactInput
) -> RusotoFuture<UpdateProvisioningArtifactOutput, UpdateProvisioningArtifactError>

Updates the specified provisioning artifact (also known as a version) for the specified product.

You cannot update a provisioning artifact for a product that was shared with you.

fn update_tag_option(
&self,
input: UpdateTagOptionInput
) -> RusotoFuture<UpdateTagOptionOutput, UpdateTagOptionError>

Updates the specified TagOption.